+++ kornia/utils/__init__.py
@@ -2,6 +2,7 @@ from .one_hot import one_hot
 from .grid import create_meshgrid, create_meshgrid3d
 from .image import tensor_to_image, image_to_tensor
 from .pointcloud_io import save_pointcloud_ply, load_pointcloud_ply
+from .helpers import _extract_device_dtype
 
 from kornia.utils.metrics import *
 
@@ -13,4 +14,5 @@ __all__ = [
     "image_to_tensor",
     "save_pointcloud_ply",
     "load_pointcloud_ply",
+    "_extract_device_dtype",
 ]
